# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Melete",
    "version": "0.1.0-pre",
    "summary": "An open-source, self-hosted, model-agnostic personal assistant.",
    "description": "Give Melete a responsibility, close the tab, come back to progress, a result, or one precise question. This document describes the v0.1 HTTP surface. Nothing here is stable yet: the release is pre-release and endpoints may change until v0.1.0 is tagged.",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8787",
      "description": "Default self-hosted address"
    }
  ],
...
  "paths": {
...
    "/knowledge": {
      "get": {
        "tags": [
          "knowledge"
        ],
        "summary": "List the records of one space",
        "description": "The catalog of the space the caller is bound to. A space id may be repeated as a query argument and must match; it never selects a different space.",
        "parameters": [
          {
            "in": "query",
            "name": "space_id",
            "schema": {
              "type": "string",
              "pattern": "^sp_[0-7][0-9A-HJKMNP-TV-Z]{25}$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Records",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "records": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "pattern": "^k_[0-7][0-9A-HJKMNP-TV-Z]{25}$"
                          },
                          "path": {
                            "type": "string"
                          },
                          "title": {
                            "type": "string"
                          },
                          "type": {
                            "$ref": "#/components/schemas/__schema114"
                          },
                          "status": {
                            "$ref": "#/components/schemas/__schema115"
                          },
                          "tags": {
                            "type": "array",
                            "items": {
                              "type": "string"
                            }
                          },
                          "updated": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "id",
                          "path",
                          "title",
                          "type",
                          "status",
                          "tags",
                          "updated"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": [
                    "records"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "403": {
            "description": "Wrong space",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/__schema58"
                }
              }
            }
          }
        }
      }
    },
    "/knowledge/search": {
      "get": {
        "tags": [
          "knowledge"
        ],
        "summary": "Search one space",
        "description": "Retrieval is scoped to exactly one space by the handle the caller holds, not by a filter argument. Retracted records are never returned.",
        "parameters": [
          {
            "in": "query",
            "name": "space_id",
            "schema": {
              "type": "string",
              "pattern": "^sp_[0-7][0-9A-HJKMNP-TV-Z]{25}$"
            },
            "required": true
          },
          {
            "in": "query",
            "name": "q",
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "required": true
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "default": 10,
              "type": "integer",
              "exclusiveMinimum": 0,
              "maximum": 50
            }
          },
          {
            "in": "query",
            "name": "include_retracted",
            "schema": {
              "default": false,
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Hits",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "hits": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "pattern": "^k_[0-7][0-9A-HJKMNP-TV-Z]{25}$"
                          },
                          "path": {
                            "type": "string"
                          },
                          "title": {
                            "type": "string"
                          },
                          "excerpt": {
                            "type": "string"
                          },
                          "status": {
                            "$ref": "#/components/schemas/__schema115"
                          },
                          "score": {
                            "type": "number"
                          }
                        },
                        "required": [
                          "id",
                          "path",
                          "title",
                          "excerpt",
                          "status",
                          "score"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": [
                    "hits"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/knowledge/{recordId}": {
      "get": {
        "tags": [
          "knowledge"
        ],
        "summary": "Read one record with its provenance",
        "parameters": [
          {
            "in": "path",
            "name": "recordId",
            "schema": {
              "type": "string",
              "description": "Knowledge record id"
            },
            "required": true,
            "description": "Knowledge record id"
          }
        ],
        "responses": {
          "200": {
            "description": "Record",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/__schema116"
                }
              }
            }
          },
          "404": {
            "description": "No such record",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/__schema58"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "knowledge"
        ],
        "summary": "Retract or delete a record",
        "parameters": [
          {
            "in": "path",
            "name": "recordId",
            "schema": {
              "type": "string",
              "description": "Knowledge record id"
            },
            "required": true,
            "description": "Knowledge record id"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "reason": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 500
                  },
                  "hard_delete": {
                    "default": false,
                    "type": "boolean"
                  }
                },
                "required": [
                  "reason"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Record",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/__schema116"
                }
              }
            }
          },
          "404": {
            "description": "No such record",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/__schema58"
                }
              }
            }
          }
        }
      }
    },
    "/knowledge/proposals": {
      "get": {
        "tags": [
          "knowledge"
        ],
        "summary": "List pending memory proposal diffs",
        "responses": {
          "200": {
            "description": "Proposals",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "proposals": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/__schema95"
                      }
                    }
                  },
                  "required": [
                    "proposals"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "knowledge"
        ],
        "summary": "Propose a knowledge write",
        "description": "The only write path an agent has. The proposal is validated and rendered as a diff the owner applies or discards; applying is a git commit.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "space": {
                    "type": "string",
                    "minLength": 1
                  },
                  "path": {
                    "type": "string",
                    "minLength": 1
                  },
                  "frontmatter": {
                    "$ref": "#/components/schemas/__schema25"
                  },
                  "body": {
                    "type": "string"
                  },
                  "rationale": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 500
                  }
                },
                "required": [
                  "space",
                  "path",
                  "frontmatter",
                  "body",
                  "rationale"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Proposal",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "proposal_id": {
                      "type": "string"
                    },
                    "path": {
                      "type": "string"
                    },
                    "diff": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "proposal_id",
                    "path",
                    "diff"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "The record failed lint",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/__schema58"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "__schema25": {
        "type": "object",
        "properties": {
          "id": {
            "$ref": "#/components/schemas/__schema26"
          },
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "space": {
            "type": "string",
            "minLength": 1
          },
          "audience": {
            "type": "string",
            "enum": [
              "private",
              "space",
              "public"
            ]
          },
          "type": {
            "type": "string",
            "enum": [
              "fact",
              "preference",
              "decision",
              "procedure",
              "reference",
              "event"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "superseded",
              "retracted",
              "disputed"
            ]
          },
          "confidence": {
            "type": "string",
            "enum": [
              "high",
              "medium",
              "low"
            ]
          },
          "asserted_by": {
            "type": "string",
            "enum": [
              "user",
              "agent",
              "document",
              "tool"
            ]
          },
          "source": {
            "type": "object",
            "properties": {
              "kind": {
                "type": "string",
                "enum": [
                  "statement",
                  "file",
                  "url",
                  "tool_output"
                ]
              },
              "ref": {
                "type": "string",
                "minLength": 1
              },
              "quote": {
                "default": "",
                "type": "string"
              },
              "sha256": {
                "default": null,
                "anyOf": [
                  {
                    "type": "string",
                    "pattern": "^[0-9a-f]{64}$"
                  },
                  {
                    "type": "null"
                  }
                ]
              }
            },
            "required": [
              "kind",
              "ref"
            ]
          },
          "observed_at": {
            "$ref": "#/components/schemas/__schema27"
          },
          "valid_from": {
            "$ref": "#/components/schemas/__schema27"
          },
          "valid_until": {
            "default": null,
            "anyOf": [
              {
                "$ref": "#/components/schemas/__schema27"
              },
              {
                "type": "null"
              }
            ]
          },
          "supersedes": {
            "default": [],
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/__schema26"
            }
          },
          "superseded_by": {
            "default": null,
            "anyOf": [
              {
                "$ref": "#/components/schemas/__schema26"
              },
              {
                "type": "null"
              }
            ]
          },
          "created": {
            "$ref": "#/components/schemas/__schema27"
          },
          "updated": {
            "$ref": "#/components/schemas/__schema27"
          },
          "tags": {
            "default": [],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "links": {
            "default": [],
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/__schema26"
            }
          },
          "schema_version": {
            "type": "number",
            "const": 1
          }
        },
        "required": [
          "id",
          "title",
          "space",
          "audience",
          "type",
          "status",
          "confidence",
          "asserted_by",
          "source",
          "observed_at",
          "valid_from",
          "created",
          "updated",
          "schema_version"
        ]
      },
      "__schema26": {
        "type": "string",
        "pattern": "^k_[0-7][0-9A-HJKMNP-TV-Z]{25}$"
      },
...
      "__schema57": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "detail": {
            "type": "object",
            "propertyNames": {
              "type": "string"
            },
            "additionalProperties": {}
          }
        },
        "required": [
          "code",
          "message"
        ],
        "additionalProperties": false
      },
      "__schema58": {
        "type": "object",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/__schema57"
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": false
      },
...
      "__schema70": {
        "type": "string",
        "minLength": 1,
        "maxLength": 240
      },
...
      "__schema95": {
        "type": "object",
        "properties": {
          "id": {
            "$ref": "#/components/schemas/__schema70"
          },
          "path": {
            "type": "string"
          },
          "diff": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "applied",
              "discarded"
            ]
          }
        },
        "required": [
          "id",
          "path",
          "diff",
          "status"
        ],
        "additionalProperties": false
      },
...
      "__schema114": {
        "type": "string",
        "enum": [
          "fact",
          "preference",
          "decision",
          "procedure",
          "reference",
          "event"
        ]
      },
      "__schema115": {
        "type": "string",
        "enum": [
          "active",
          "superseded",
          "retracted",
          "disputed"
        ]
      },
      "__schema116": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "pattern": "^k_[0-7][0-9A-HJKMNP-TV-Z]{25}$"
          },
          "path": {
            "type": "string"
          },
          "frontmatter": {
            "type": "object",
            "properties": {
              "id": {
                "$ref": "#/components/schemas/__schema117"
              },
              "title": {
                "type": "string",
                "minLength": 1,
                "maxLength": 200
              },
              "space": {
                "type": "string",
                "minLength": 1
              },
              "audience": {
                "type": "string",
                "enum": [
                  "private",
                  "space",
                  "public"
                ]
              },
              "type": {
                "$ref": "#/components/schemas/__schema114"
              },
              "status": {
                "$ref": "#/components/schemas/__schema115"
              },
              "confidence": {
                "type": "string",
                "enum": [
                  "high",
                  "medium",
                  "low"
                ]
              },
              "asserted_by": {
                "type": "string",
                "enum": [
                  "user",
                  "agent",
                  "document",
                  "tool"
                ]
              },
              "source": {
                "type": "object",
                "properties": {
                  "kind": {
                    "type": "string",
                    "enum": [
                      "statement",
                      "file",
                      "url",
                      "tool_output"
                    ]
                  },
                  "ref": {
                    "type": "string",
                    "minLength": 1
                  },
                  "quote": {
                    "default": "",
                    "type": "string"
                  },
                  "sha256": {
                    "default": null,
                    "anyOf": [
                      {
                        "type": "string",
                        "pattern": "^[0-9a-f]{64}$"
                      },
                      {
                        "type": "null"
                      }
                    ]
                  }
                },
                "required": [
                  "kind",
                  "ref",
                  "quote",
                  "sha256"
                ],
                "additionalProperties": false
              },
              "observed_at": {
                "$ref": "#/components/schemas/__schema118"
              },
              "valid_from": {
                "$ref": "#/components/schemas/__schema118"
              },
              "valid_until": {
                "default": null,
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/__schema118"
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "supersedes": {
                "default": [],
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/__schema117"
                }
              },
              "superseded_by": {
                "default": null,
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/__schema117"
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "created": {
                "$ref": "#/components/schemas/__schema118"
              },
              "updated": {
                "$ref": "#/components/schemas/__schema118"
              },
              "tags": {
                "default": [],
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "links": {
                "default": [],
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/__schema117"
                }
              },
              "schema_version": {
                "type": "number",
                "const": 1
              }
            },
            "required": [
              "id",
              "title",
              "space",
              "audience",
              "type",
              "status",
              "confidence",
              "asserted_by",
              "source",
              "observed_at",
              "valid_from",
              "valid_until",
              "supersedes",
              "superseded_by",
              "created",
              "updated",
              "tags",
              "links",
              "schema_version"
            ],
            "additionalProperties": false
          },
          "body": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "path",
          "frontmatter",
          "body"
        ],
        "additionalProperties": false
      },
      "__schema117": {
        "type": "string",
        "pattern": "^k_[0-7][0-9A-HJKMNP-TV-Z]{25}$"
      },
      "__schema118": {
        "type": "string",
        "format": "date",
        "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))$"
      },
...
    },
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "melete_session"
      }
    }
  }
}